Use MUnit in HListToFuncSuite
